Repo Data Science for All - Spring 2025
Materials for CS118/MATH118 SBCC Spring 2025: This repo contains the publicly available materials that are used in the Data Science for All course (CS/MATH 118). This includes: homeworks, labs, and notebooks used in lecture. The contents of this repository are licensed for reuse under Creative Commons Attribution-NonCommercial 4.0 International (CC BY-NC 4.0) (some of the materials are based on Data 8 at UCB: data8.org)